Cleanup, and make web UI run first.
