+++ node/src/web_common.rs
@@ -1,0 +1,31 @@
+use axum::body::Body;
+use axum::http::{Response, StatusCode};
+use axum::response::IntoResponse;
+use prometheus::{default_registry, Encoder, TextEncoder};
+
+/// Wrapper to make Axum understand how to convert anyhow::Error into a 500
+/// response.
+pub(crate) struct AnyhowErrorWrapper(anyhow::Error);
+
+impl From<anyhow::Error> for AnyhowErrorWrapper {
+    fn from(e: anyhow::Error) -> Self {
+        AnyhowErrorWrapper(e)
+    }
+}
+
+impl IntoResponse for AnyhowErrorWrapper {
+    fn into_response(self) -> Response<Body> {
+        Response::builder()
+            .status(StatusCode::INTERNAL_SERVER_ERROR)
+            .body(Body::from(format!("{:?}", self.0)))
+            .unwrap()
+    }
+}
+
+pub(crate) async fn metrics() -> String {
+    let metric_families = default_registry().gather();
+    let mut buffer = vec![];
+    let encoder = TextEncoder::new();
+    encoder.encode(&metric_families, &mut buffer).unwrap();
+    String::from_utf8(buffer).unwrap()
+}
